Removed the last dependencies breaking no-std build. (#669)
* Removed the last dependencies breaking no-std build.

Removed to_binary, which was used only in a debug display.

Replaced once_cell with lazy_static. The no-std build of once_cell did
not support the Sync version of OnceCell, that was needed when creating
static variables. On the other hand, lazy_static is the complete
solution for lazyly initialized (Sync) static variables, and is no-std.

With this changes, revm builds on a no-std platform.

* Reverting back to OnceCell.

* Hex encoding debug print.

* Using race::OnceBox instead of critical-section.

* Lint error.