Never inline the prepare functions
These functions are used for keeping the stack memory clean on the
host recursive code paths. EVM supports up to 1024 levels of recursion,
so if one is not careful with the stack memory allocations, the stack
memory can blow up.

Benchmarks show dramatic stack memory improvements when ont inlining
these functions. A recursive bomb uses around 1.1MB of stack when
these functions are not inlined. If inlined the recursive bombs blow
up the stack.