feat(context): add mark_cold method to JournaledAccount (#3160)
* feat(context): add mark_cold method to JournaledAccount

Add a public `mark_cold()` method to `JournaledAccount` that delegates
to the underlying `Account::mark_cold()`. This is analogous and provides
a way to mark accounts as cold that was available in previous revm
versions.

This is used by celo-revm to provide backwards compatibility for the
transfer precompile, which does not warm addresses in its initial
implementation.

* Declare mark_cold to be unsafe

and add explanation.

Co-authored-by: rakita <rakita@users.noreply.github.com>

---------

Co-authored-by: rakita <rakita@users.noreply.github.com>